Update README.md
The Project devoted to the analysis of the A/B Test performed in the startup that sells food products. We investigated user behavior for the company's app. First step will be the study of the sales funnel: what number of users reaches the purchase stage and how many users actually make it to this stage, how many get stuck at previous stages. Also we will investigate on which stages in particular the biggest loss of users occurs.
The second step: analyzing the results of an A/A/B test. The designers would like to change the fonts for the entire app, but the managers are afraid the users might find the new design intimidating. The decision whether implement changes or not will be made based on the results of an A/A/B test. The users are split into three groups: two control groups get the old fonts and one test group gets the new ones.
Creating two A groups has certain advantages. We can make it a principle that we will only be confident in the accuracy of our testing when the two control groups are similar. If there are significant differences between the A groups, this can help us uncover factors that may be distorting the results. Comparing control groups also tells us how much time and data we'll need when running further tests.